pyproject.toml
[project]
name = "image-processing"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.68",
    "langchain-google-genai>=2.1.7",
    "langgraph>=0.5.3",
    "pillow>=11.3.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.1",
    "streamlit>=1.46.1",
]
